requirements.txt
# Core dependencies
fidelity-api==0.0.16
python-dotenv==1.0.1
yfinance==0.2.52

# CLI framework
click==8.1.7
rich==13.7.0

# Configuration
pyyaml==6.0.1
keyring==24.3.0

# Logging
loguru==0.7.2

# Scheduler
apscheduler==3.10.4

# Web dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
